+++ packages/babel-preset-expo/src/plugins/define-plugin.ts
