+++ packages/babel-preset-expo/src/plugins/detect-dynamic-exports.ts
