+++ packages/babel-preset-expo/src/plugins/minify-platform-select-plugin.ts
