+++ packages/babel-preset-expo/src/plugins/restricted-react-api-plugin.ts
