+++ packages/babel-preset-expo/src/plugins/server-metadata-plugin.ts
@@ -6,7 +6,7 @@
  */
 import type { ConfigAPI, types as t, PluginObj, PluginPass } from '@babel/core';
 
-import { getExpoRouterAbsoluteAppRoot, getIsServer, toPosixPath } from './common';
+import { getExpoRouterAbsoluteAppRoot, getIsServer, toPosixPath } from '../common';
 
 const debug = require('debug')('expo:babel:server-metadata');
